🚧 (webapp) Fix load on homepage
