+++ packages/web/webapp-svelte/src/composition/useLoad.ts
@@ -1,22 +1,22 @@
 import type { LoadInput } from "@sveltejs/kit";
-import type { IterableElement } from "type-fest";
-import { loadSession } from "./useSession";
+import { loadSession, patchSession } from "./useSession";
 
-export function useLoad<T extends Array<(...args: unknown[]) => unknown>>(
+type ReturnTypes<T> = T extends [...infer U, infer A]
+  ? A extends (...args: unknown[]) => unknown
+    ? ReturnType<A> & ReturnTypes<U>
+    : void
+  : void;
+
+export function useLoad<T extends (...args: unknown[]) => unknown>(input: LoadInput, fn: T): ReturnType<T>;
+export function useLoad<T1 extends (...args: unknown[]) => unknown, T2 extends (...args: unknown[]) => unknown>(
   input: LoadInput,
-  ...fns: T
-): ReturnType<IterableElement<T>> {
+  fn1: T1,
+  fn2: T2
+): ReturnType<T1> & ReturnType<T2>;
+export function useLoad<T extends Array<(...args: unknown[]) => unknown>>(input: LoadInput, ...fns: T): ReturnTypes<T> {
   input.session.fetch = input.fetch;
 
-  input.session.toJSON = function (this) {
-    const clone = { ...this };
-
-    delete clone.toJSON;
-    delete clone.fetch;
-    delete clone.stores;
-
-    return clone;
-  };
+  patchSession(input.session);
 
   loadSession.set(input.session);
